requirements.txt
fastapi
uvicorn
pillow
rembg==2.0.67
onnxruntime
numpy
opencv-python-headless
python-multipart
